### 2018-10-16 #261 @juyeongkim
Submission: DataSpaceR
DataSpaceR is an R interface to the CAVD DataSpace, a data sharing and discovery tool that facilitates exploration of HIV immunological data from pre-clinical and clinical HIV vaccine studies.

```
Package: DataSpaceR
Title: An R Interface to the CAVD DataSpace
Version: 0.5.2
Authors@R: c(person("Ju Yeong", "Kim",
                    email = "jkim2345@fredhutch.org",
                    role = c("aut", "cre")),
             person("The Collaboration for AIDS Vaccine Discovery",
                    role = "cph"))
Description: Provides a convenient API for accessing datasets within
    the CAVD DataSpace (dataspace.cavd.org).
License: GPL-3 + file LICENSE
    R6,
    Rlabkey (>= 2.2.0),
    curl,
    httr,
    assertthat,
    digest,
    rjson,
    data.table
    covr,
    pryr
```

https://github.com/CAVDDataSpace/DataSpaceR

[e.g., "data extraction, because the package parses a scientific data file format"]

The package fits into the **data retrieval** category as it provides API access to the database.

The target audience are immunologists, bioinformaticians, or statisticians in HIV vaccine research, and the scientific applications are meta analyses of exploring relationships across assays, studies, and time.

None that I am aware of.

https://github.com/ropensci/onboarding/issues/260

  - I've used camelCase in this package.
  - The test coverage is currently 92%, and I am getting LaTex errors with `goodpractice`:
```R
── GP DataSpaceR ────────────────────────────────────────────────────────────────────────────────

  ✖ write unit tests for all functions, and all package code in general. 92%
    of code lines are covered by test cases.

    R/DataSpaceConnection.R:161:NA
    R/helpers.R:29:NA
    R/helpers.R:49:NA
    R/helpers.R:52:NA
    R/helpers.R:152:NA

  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version.
    This typically indicates Rd problems. LaTeX errors found: ! LaTeX Error: File
    `inconsolata.sty' not found. Type X to quit or <RETURN> to proceed, or enter new
    name. (Default extension: sty)
─────────────────────────────────────────────────────────────────────────────────────────────────
```

@seaaan
