### 2018-10-24 #261 @noamross
### Editor checks:

---

#### Editor comments

Thanks for your submission @juyeongkim and sorry for being a little slow on the pick-up.  What a neat package design!  Unfortunately I'm getting a few errors while testing.  In R CMD Check, I get:

```
* creating vignettes ... ERROR
Quitting from lines 146-148 (Intro_to_DataSpaceR.Rmd)
Error: processing vignette 'Intro_to_DataSpaceR.Rmd' failed with diagnostics:
216 is not a valid group ID
Execution halted
Error: Command failed (1)
Execution halted
```

And running `test_package`, I get:

```
Loading DataSpaceR
Loading required package: testthat
By exporting data from the CAVD DataSpace, you agree to be bound by the Terms of Use available on the CAVD DataSpace sign-in page at https://dataspace.cavd.org/cds/CAVD/app.view?
Testing DataSpaceR
✔ | OK F W S | Context
✔ | 40       | DataSpaceConnection [7.6 s]
✖ | 19 1     | helpers [1.1 s]
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
test-helpers.R:21: failure: `getUserEmail`
DataSpaceR:::getUserEmail(url, NULL) not equal to `email`.
1/1 mismatches
x[1]: "noam.ross@gmail.com"
y[1]: "jkim2345@scharp.org"
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |  9       | netrc
✔ | 73       | DataSpaceStudy (CAVD) [191.7 s]
✔ | 64       | DataSpaceStudy (cvd408) [9.2 s]
✖ |  0 2     | DataSpaceStudy (mice)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
test-study.R:26: failure: can connect to studies
`cavd` inherits from `try-error` not `DataSpaceStudy`.

test-study.R:27: failure: can connect to studies
`cavd` inherits from `try-error` not `R6`.
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ |  0 2     | DataSpaceStudy (CAVD 242)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
test-study.R:26: failure: can connect to studies
`cavd` inherits from `try-error` not `DataSpaceStudy`.

test-study.R:27: failure: can connect to studies
`cavd` inherits from `try-error` not `R6`.
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔ | 73       | DataSpaceStudy (NYVAC durability comparison) [67.8 s]

══ Results ═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Duration: 280.0 s

OK:       278
Failed:   5
Warnings: 0
Skipped:  0
```

I note that in the neither in web GUI nor in `con$availableGroups` do I see group 216.  I suspect that I don't have the same permissions you do and that your tests and vignettes attempting to access studies not available to all users. (I see the first test failure is because you've hard-coded your email into the tests, so it doesn't work once I've made my own netrc file). If this is the case, you should probably either select a different subset of studies, or skip tests conditionally based on user permissions.

Everything else checks out fairly well, though, so I'm happy to start seeking reviewers while you patch this.
---

Reviewers:
Due date:
