### 2018-10-26 #261 @juyeongkim
Thanks @noamross! I will fix these tests today.
