### 2018-10-26 a93b777
Skip testing private groups for others
https://github.com/ropensci/onboarding/issues/261#issuecomment-432830937
