### 2018-11-26 #261 @noamross
Thanks for your thorough review, @seaaan! @juyeongkim, I suggest waiting for @czibman's review so you can address comments together.
