### 2018-11-27 #261 @czibman
DataSpaceR review
================

## Package Review

*Please check off boxes as applicable, and elaborate in comments below.
Your review is not limited to these topics, as described in the reviewer
guide*

  - \[x\] As the reviewer I confirm that there are no conflicts of
    interest for me to review this work (If you are unsure whether you
    are in conflict, please speak to your editor *before* starting your
    review).

#### Documentation

  - \[x\] **A statement of need** clearly stating problems the software
    is designed to solve and its target audience in README
  - \[x\] **Installation instructions:** for the development version of
    package and any non-standard dependencies in README
  - \[x\] **Vignette(s)** demonstrating major functionality that runs
    successfully locally
  - \[x\] **Function Documentation:** for all exported functions in R
    help
  - \[x\] **Examples** for all exported functions in R Help that run
    successfully locally
  - \[x\] **Community guidelines** including contribution guidelines in
    the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports`
    and `Maintainer` (which may be autogenerated via `Authors@R`).

##### For packages co-submitting to JOSS

N/A

#### Functionality

  - \[x\] **Installation:** Installation succeeds as documented.
  - \[x\] **Functionality:** Any functional claims of the software been
    confirmed.
  - \[x\] **Performance:** Any performance claims of the software been
    confirmed.
  - \[x\] **Automated tests:** Unit tests cover essential functions of
    the package and a reasonable range of inputs and conditions. All
    tests pass on the local machine.
  - \[x\] **Packaging guidelines**: The package conforms to the rOpenSci
    packaging guidelines

#### Final approval (post-review)

  - \[ \] **The author has responded to my review and made changes to my
    satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: 8 —

### Review Comments

#### Summary

Thank you so much for making these data more accessible to researchers
This appears to be a well-designed package. I note below where some
improvements can be make the use of this package more dummy-proof.

#### README

Per ROpenSci’s requirement that the Statement of Need specify the user
population, consider adding a line or two about who a typical user of
the CAVD data might be.

Please also consider more explicitly describing contribution guidelines
in the README (in keeping with the “Community Guidelines” Requirement).

#### Installation

I had to use “force = TRUE” to get the package to
    load.

``` r
devtools::install_github("CAVDDataSpace/DataSpaceR")
```

    ##   Use `force = TRUE` to force installation

#### Creation of CAVD account

The package creator instructs the user for a CAVD account at
<https://dataspace.cavd.org>. I recognize that the registration process
is beyond the influence of the package creator, but registering for a
CAVD account was tricky since, when I clicked “submit” on the “member
details” page, nothing happened. Submitting member details appears
optional, however.

#### Vignette

Note that, when loading the package, a statement appears reminding the
user of the Terms of Use. That’s fine, but it ends in a question mark
(that is part of the url name). Makes me think I’m being prompted for
something. This can be reworded as “see \[url\] for the
    terms”.

``` r
library(DataSpaceR)
```

    ## By exporting data from the CAVD DataSpace, you agree to be bound by the Terms of Use available on the CAVD DataSpace sign-in page at https://dataspace.cavd.org/cds/CAVD/app.view?

The Vignette has the reader look at the list of available groups and
choose group 216. There is no group 216 (see below). Is this because the
group was deleted from the CAVD dataset? Is there a way to ensure that
the documentation for this package stays current when the dataset
changes?

``` r
knitr::kable(con$availableGroups)
```

|  id | label                              | originalLabel                      | description                                                                                                               | createdBy | shared |   n | studies                                   |
| --: | :--------------------------------- | :--------------------------------- | :------------------------------------------------------------------------------------------------------------------------ | :-------- | :----- | --: | :---------------------------------------- |
| 220 | NYVAC durability comparison        | NYVAC\_durability                  | Compare durability in 4 NHP studies using NYVAC-C (vP2010) and NYVAC-KC-gp140 (ZM96) products.                            | ehenrich  | TRUE   |  78 | c(“cvd281”, “cvd434”, “cvd259”, “cvd277”) |
| 228 | HVTN 505 case control subjects     | HVTN 505 case control subjects     | Participants from HVTN 505 included in the case-control analysis                                                          | drienna   | TRUE   | 189 | vtn505                                    |
| 230 | HVTN 505 polyfunctionality vs BAMA | HVTN 505 polyfunctionality vs BAMA | Compares ICS polyfunctionality (CD8+, Any Env) to BAMA mfi-delta (single Env antigen) in the HVTN 505 case control cohort | drienna   | TRUE   | 170 | vtn505                                    |

I was able to follow the rest of the vignette without trouble. As a very
minor quibble, the vignette reference at least one other package
(`pryr`). To allow for a faster run through the vignette that can be
done entirely by cutting and pasting, it would be helpful to add code in
the vignette to install any referenced packages if the user has not
already (as in below).

``` r
if (!("pryr" %in% installed.packages())) install.packages("pryr")
```

#### Documentation and testing

The package includes 3 functions: writeNetrc, checkNetrc, and connectDS.
Help files are provided for:

  - DataSpaceConnection
  - DataSpaceR-package
  - DataSpaceStudy
  - checkNetrc
  - connectDS
  - writeNetrc

The “DataSpaceR-package” help file does not appear to be loading
correctly when I type `?DataSpaceR-package`(I get the help file for
Arithmetic Operators instead).

All three function help files include examples. The information in the
README and vignette on creating a .netrc file is helpful. The help file
for the `writeNetrc` function, in contrast, could benefit from
containing more of the information that is in the vignette (for example
on how to create and save the file). In addition, it would be helpful if
the `writeNetrc` command provided the option to overwrite an existing
netrc file (finding and changing a file with a non-standard name was
harder than it should have been on my mac).

There are four files containing tests:

  - test-connection.R

  - test-helpers.R

  - test-netrc.R

  - test-study.R

#### Additional comments:

It’s not clear to me what the purpose of a group is. I would have
thought it was a group of studies, but it is a group of datasets.

When you connect, you are provided with information regarding the number
of studies, subjects, assays, datapoints, and groups in the CAVD
database. It would be helpful if the documentation explained the role of
the five assays and why the number of assays is a meaningful number to
report (unless this information is obvious to people who’d use this
data). Perhaps also report the number of datasets.
