### 2018-12-13 af8a2ba
Update the link to the sign-in page
