### 2018-12-13 de9a65d
Show how one can create a netrc file
