### 2018-12-14 3c46588
Reexport Rlabkey::makeFilter
