+++ DESCRIPTION
@@ -30,4 +30,4 @@ Suggests:
     knitr,
     pryr
 VignetteBuilder: knitr
-RoxygenNote: 6.1.0
+RoxygenNote: 6.1.1
