### 2018-12-14 996c414
Update class documentation
