### 2018-12-14 5b74be2
Check if netrc exists before checking credential
