### 2018-12-14 507e21c
Move netrc check to initalization in connection object and document
