+++ R/netrc.R
@@ -2,6 +2,13 @@
 #'
 #' @description Write a netrc file that is valid for accessing DataSpace.
 #'
+#' @details
+#' The database is accessed with the user's credentials.
+#' A netrc file storing login and password information is required.
+#' See \href{https://cavddataspace.github.io/DataSpaceR/}{here}
+#' for instruction on how to register and set DataSpace credential.
+#' By default \code{curl} will look for the file in your home directoty.
+#'
 #' @param login A character. Email address used for logging in on DataSpace.
 #' @param password A character. Password associated with the login.
 #' @param onStaging A logical. Whether to connect to the staging server instead
@@ -13,7 +20,9 @@
 #' @seealso \code{\link{connectDS}} \code{\link{checkNetrc}}
 #' @examples
 #' \dontrun{
-#' writeNetrc("dataspaceuser@email.com", "mypassword")
+#' # First, create an account in the DataSpace App and read the terms of use
+#' # Next, create a netrc file using writeNetrc()
+#' writeNetrc("dataspaceuser@email.com", "yourSecretPassword")
 #' }
 #' @export
 writeNetrc <- function(login,
@@ -49,21 +58,11 @@ writeNetrc <- function(login,
 #'
 #' @description Check that there is a netrc file with a valid entry for DataSpace.
 #'
+#' @param netrcFile A character. File path to netrc file to check.
 #' @param onStaging A logical. Whether to check the staging server instead
 #' of the production server.
-#'
-#' @details
-#' In order to connect to DataSpace, you will need a \code{.netrc} file in your
-#' contains a \code{machine} name (hostname of DataSpace), and \code{login} and
-#' \code{password}.
-#' See \href{https://www.labkey.org/wiki/home/Documentation/page.view?name=netrc}{here}
-#' for more information. By default \code{RCurl} will look for the file in your
-#' home directoty.
-#'
-#' If no netrc is available or it is not formatted properly,
-#' \code{\link{writeNetrc}}  can be used to write one.
-#' Otherwise, when specifying login and password in \code{connectDS},
-#' a temporary file will be created for that connection.
+#' @param verbose A logical. Whether to print the extra details for
+#' troubleshooting.
 #'
 #' @return The name of the netrc file
 #' @seealso \code{\link{connectDS}} \code{\link{writeNetrc}}
@@ -72,11 +71,11 @@ writeNetrc <- function(login,
 #' checkNetrc()
 #' }
 #' @export
-checkNetrc <- function(onStaging = FALSE, verbose = TRUE) {
+checkNetrc <- function(netrcFile = getNetrcPath(),
+                       onStaging = FALSE,
+                       verbose = TRUE) {
   if (exists("labkey.netrc.file", .GlobalEnv)) {
     netrcFile <- get("labkey.netrc.file", .GlobalEnv)
-  } else {
-    netrcFile <- getNetrcPath()
   }
 
   if (!file.exists(netrcFile)) {
@@ -97,7 +96,7 @@ checkNetrc <- function(onStaging = FALSE, verbose = TRUE) {
   }
 
   if (verbose) {
-  message("netrc file found at '", netrcFile, "', and it looks valid.")
+    message("netrc file found at '", netrcFile, "', and it looks valid.")
   }
 
   netrcFile
