### 2018-12-18 cd35008
Rename getVariableInfo to getDatasetDescription
