### 2018-12-18 2b466c8
Rename getVariableInfo to getDatasetDescription
