### 2018-12-18 02e1032
Update vignette with Sean's suggestions
