+++ vignettes/Intro_to_DataSpaceR.Rmd
@@ -60,6 +60,12 @@ con
 
 The call to `connectDS` instantiates the connection. Printing the object shows where it's connected and the available studies.
 
+```{r availableStudies}
+knitr::kable(head(con$availableStudies))
+```
+
+`con$availableStudies` shows the available studies in the CAVD DataSpace.
+
 ```{r getStudy}
 cvd256 <- con$getStudy("cvd256")
 cvd256
@@ -87,10 +93,10 @@ colnames(NAb)
 
 The *cvd256* object is an [`R6`](https://cran.r-project.org/web/packages/R6/) class, so it behaves like a true object. Functions (like `getDataset`) are members of the object, thus the `$` semantics to access member functions.
 
-We can get detailed variable information using `getVariableInfo`.
+We can get detailed variable information using `getDatasetDescription`.
 
 ```{r getVariableInfo}
-knitr::kable(cvd256$getVariableInfo("NAb"))
+knitr::kable(cvd256$getDatasetDescription("NAb"))
 ```
 
 To get only a subset of the data and speed up the download, filters can be passed to `getDataset`. The filters are created using the `makeFilter` function of the `Rlabkey` package.
@@ -104,11 +110,11 @@ dim(NAb_day0)
 See `?makeFilter` for more information on the syntax.
 
 
-## Cross-study connection
+## Creating a connection to all studies
 
 To fetch data from multiple studies, create a connection at the project level.
 
-```{r cross-connection}
+```{r connection-all-studies}
 cavd <- con$getStudy("")
 ```
 
@@ -116,14 +122,14 @@ This will instantiate a connection at the `CAVD` level. Most functions work cros
 
 You can get a list of datasets available across all studies.
 
-```{r cross-connection-print}
+```{r connection-all-studies-datasets}
 cavd
 knitr::kable(cavd$availableDatasets)
 ```
 
-In cross-study connections, `getDataset` will combine the requested datasets. Note that in most cases the datasets will have too many subjects for quick data transfer, making filtering of the data a necessity. The `colFilter` argument can be used here, as described in the `getDataset` section.
+In all-study connection, `getDataset` will combine the requested datasets. Note that in most cases, the datasets will have too many subjects for quick data transfer, making filtering of the data a necessity. The `colFilter` argument can be used here, as described in the `getDataset` section.
 
-```{r cross-connection-dem}
+```{r connection-all-studies-dem}
 conFilter <- makeFilter(c("species", "EQUAL", "Human"))
 human <- cavd$getDataset("Demographics", colFilter = conFilter)
 dim(human)
@@ -133,69 +139,63 @@ colnames(human)
 
 ## Connect to a saved group
 
-We can browse available saved groups via `availableGroups`.
+A group is a curated collection of participants from filtering of treatments, products, studies, or species, and it is created in [the DataSpace App](https://dataspace.cavd.org/cds/CAVD/app.view).
+
+Let's say you are using the App to filter and visualize data and want to save them for later or explore in R with `DataSpaceR`. You can save a group by clicking the Save button on the Active Filter Panel.
+
+We can browse available the saved groups or the curated groups by DataSpace Team via `availableGroups`.
 
 ```{r availableGroups}
 knitr::kable(con$availableGroups)
 ```
 
-To fetch data from a saved group, create a connection at the project level with a group ID. For example, we can connect to `mice` group which has group ID 216 by `getGroup`.
[…]
