### 2018-12-18 68f3e06
Do not run tests or build vignettes in CRAN
Since I cannot set credentials in CRAN, I need to disable running tests or building vignettes in CRAN.

Read this for more info: https://cran.r-project.org/web/packages/googlesheets/vignettes/managing-auth-tokens.html#the-making-of-googlesheets.tar.gz-for-cran
