### 2018-12-18 2fcfded
Deprecate groupId arugment in getStudy
