### 2018-12-18 4dae8b9
Use getFromNamespace instead of triple colon :::
