+++ NAMESPACE
@@ -27,4 +27,5 @@ importFrom(digest,digest)
 importFrom(httr,GET)
 importFrom(httr,content)
 importFrom(rjson,fromJSON)
+importFrom(utils,getFromNamespace)
 importFrom(utils,packageVersion)
