### 2018-12-19 a8a3713
Add more info to README
