### 2018-12-19 81d1f34
Update writeNetrc unit test expectation
