### 2018-12-19 73b6f92
Remove number of asssays print statement
