+++ R/DataSpaceConnection.R
@@ -133,7 +133,6 @@ DataSpaceConnection <- R6Class(
       cat("\n  Available studies:", private$.stats$studies)
       cat("\n    -", private$.stats$subjectlevelstudies, "studies with data")
       cat("\n    -", private$.stats$subjects, "subjects")
-      cat("\n    -", private$.stats$assays, "assays")
       cat("\n    -", private$.stats$datacount, "data points")
       cat("\n  Available groups:", nrow(private$.availableGroups))
       cat("\n")
