+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: DataSpaceR
 Type: Package
 Title: An R Interface to the CAVD DataSpace
-Version: 0.5.2
+Version: 0.6.0
 Authors@R: c(person("Ju Yeong", "Kim",
                     email = "jkim2345@fredhutch.org",
                     role = c("aut", "cre")),
