### 2018-12-19 #261 @juyeongkim
Thank you @seaaan and @czibman for reviewing the package!
