### 2018-12-19 51ed1d2
Update appveyor.yml
