### 2018-12-19 44f901b
Update appveyor.yml
