+++ appveyor.yml
@@ -45,6 +45,3 @@ artifacts:
 
   - path: '\*_*.zip'
     name: Bits
-
-on_finish:
-  - ps: $blockRdp = $true; iex ((new-object net.webclient).DownloadString('https://raw.githubusercontent.com/appveyor/ci/master/scripts/enable-rdp.ps1'))
