### 2018-12-20 deb45c9
Skip clear_cache unit test if pryr is not installed
