### 2018-12-28 #261 @seaaan
Nice changes! Sorry it took me a bit to get a chance to review them. The updated package builds fine and passes R CMD CHECK. I have a few minor comments--and I do encourage you to consult with users about changing the interface as discussed above--but other than those, I don't have anything else and I approve the package. Nice job!

## Vignette: how to get help
I suggest adding a short section in the vignette (and possibly README) explaining how to access documentation, since the help for R6 classes works a bit differently than for most of R.

## Reference tables
For the reference tables of the vignette, add a little explanatory text explaining what the reference tables are meant to be used for.

Also, I suggest dropping a number of little-used methods from the tables. This will make the tables more useful for finding the most commonly used functions. Specifically, for `DataSpaceConnection`: drop `config`, `initialize`, `print`, and possibly `refresh`; for `DataSpaceStudy`: drop `study`, `group`, `config`, `cache`, `clearCache`, `initialize`, `print`, and possibly `refresh`.

I don't think you need the "type" column in these tables.
