### 2018-12-31 b1ec75a
Revise reference tables and add links to reference pages
