+++ README.md
@@ -1,33 +1,52 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-DataSpaceR <img src="man/figures/logo.png" align="right" />
-===========================================================
 
-[![Build Status](https://travis-ci.org/CAVDDataSpace/DataSpaceR.svg?branch=master)](https://travis-ci.org/CAVDDataSpace/DataSpaceR) [![Build status](https://ci.appveyor.com/api/projects/status/bmwyv5i32xr07bdr/branch/master?svg=true)](https://ci.appveyor.com/project/juyeongkim/dataspacer/branch/master) [![codecov](https://codecov.io/gh/CAVDDataSpace/DataSpaceR/branch/master/graph/badge.svg)](https://codecov.io/gh/CAVDDataSpace/DataSpaceR/branch/master)
+# DataSpaceR <img src="man/figures/logo.png" align="right" />
 
-`DataSpaceR` is an R interface to [the CAVD DataSpace](https://dataspace.cavd.org), a data sharing and discovery tool that facilitates exploration of HIV immunological data from pre-clinical and clinical HIV vaccine studies.
+[![Build
+Status](https://travis-ci.org/CAVDDataSpace/DataSpaceR.svg?branch=master)](https://travis-ci.org/CAVDDataSpace/DataSpaceR)
+[![Build
+status](https://ci.appveyor.com/api/projects/status/bmwyv5i32xr07bdr/branch/master?svg=true)](https://ci.appveyor.com/project/juyeongkim/dataspacer/branch/master)
+[![codecov](https://codecov.io/gh/CAVDDataSpace/DataSpaceR/branch/master/graph/badge.svg)](https://codecov.io/gh/CAVDDataSpace/DataSpaceR/branch/master)
 
-The package is intended for use by immunologists, bioinformaticians, and statisticians in HIV vaccine research, or anyone interested in the analysis of HIV immunological data across assays, studies, and time.
+`DataSpaceR` is an R interface to [the CAVD
+DataSpace](https://dataspace.cavd.org), a data sharing and discovery
+tool that facilitates exploration of HIV immunological data from
+pre-clinical and clinical HIV vaccine studies.
 
-This package simplifies access to the database by taking advantage of the standardization of the database to hide all the [Rlabkey](https://cran.r-project.org/web/packages/Rlabkey/index.html) specific code away from the user, and it allows the users to access the study-specific datasets via [an object-oriented paradigm](https://cran.r-project.org/web/packages/R6/readme/README.html).
+The package is intended for use by immunologists, bioinformaticians, and
+statisticians in HIV vaccine research, or anyone interested in the
+analysis of HIV immunological data across assays, studies, and time.
 
-Installation
-------------
+This package simplifies access to the database by taking advantage of
+the standardization of the database to hide all the
+[Rlabkey](https://cran.r-project.org/web/packages/Rlabkey/index.html)
+specific code away from the user, and it allows the users to access the
+study-specific datasets via [an object-oriented
+paradigm](https://cran.r-project.org/web/packages/R6/readme/README.html).
 
-You can install the latest version of DataSpaceR from [GitHub](https://github.com/CAVDDataSpace/DataSpaceR) with [devtools](https://cran.r-project.org/web/packages/devtools/index.html):
+## Installation
+
+You can install the latest version of DataSpaceR from
+[GitHub](https://github.com/CAVDDataSpace/DataSpaceR) with
+[devtools](https://cran.r-project.org/web/packages/devtools/index.html):
 
 ``` r
 # install.packages("devtools")
 devtools::install_github("CAVDDataSpace/DataSpaceR")
 ```
 
-Register and set DataSpace credential
--------------------------------------
+## Register and set DataSpace credential
 
[…]
