### 2018-12-31 #261 @juyeongkim
@seaaan Thanks! I made the changes [here](https://github.com/CAVDDataSpace/DataSpaceR/commit/b1ec75a6e1ccfce5f50f454b392a82312f4c8941).

Happy new year! :fireworks: :champagne:
