### 2019-01-03 #261 @czibman
Apologies for taking so long to look at this.  I don't have any further comments.  Thanks @juyeongkim again for making this package, and happy new year to you all as well!
