### 2019-01-31 98ec0e4
Finalize rOpenSci submission
* Add rOpenSci footer to README
* Fix links
* Add codemeta.json
