+++ NEWS.md
@@ -1,3 +1,7 @@
+# DataSpaceR 0.6.0
+
+* rOpenSci submission: https://github.com/ropensci/software-review/issues/261
+
 # DataSpaceR 0.5.2
 
 * Modified `getDataset` method in `DataSpaceStudy` to take `mergeExtra` as an argument in order to merge extra information (demographics and treatment arm). #5
