### 2019-01-31 #261 @juyeongkim
@noamross Awesome! 🍾

I've checked off everything in the to-do list:
[x] Transfer the repo to rOpenSci
[x] Add the rOpenSci header to README
[x] Fix links
[x] Add codemeta.json

I'd like to acknowledge my reviewers in my package DESCRIPTION and/or README. Would you like to be acknowledged, @seaaan and @czibman?

I would love to write a blog post about my package. (I still need to write blog posts on RSelenium...)

Thank you @noamross, @seaaan, and @czibman for the review and feedback. It's been undoubtedly productive. 💯
