### 2019-01-31 #261 @stefaniebutland
> I would love to write a blog post about my package

woohoo!
- Blog posts by authors of peer-reviewed packages so you can get an idea of narrative style and length you prefer: https://ropensci.org/tags/software-peer-review/
- Tech Notes: https://ropensci.org/technotes/
- Technical and editorial guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post

Give me an idea of when you would like to submit a draft via pull request. Don't worry about the RSelenium posts - would love to have one but there's no obligation and you could do one any time it works for you.

Happy to answer any questions @juyeongkim.
