### 2019-02-02 b941fdd
Credit Sean as reviewer and add badges
