+++ .trigger.sh
@@ -1,3 +1,5 @@
+if [ "$TRAVIS_OS_NAME" == "linux"]
+then
 body='{
   "request": {
     "branch": "master",
@@ -12,3 +14,5 @@ curl -s -X POST \
   -H "Authorization: token $TRAVIS_API_TOKEN" \
   -d "$body" \
   https://api.travis-ci.org/repo/CAVDDataSpace%2FDataSpaceR.test/requests
+fi
+
