### 2019-02-02 58e70ec
Update .trigger.sh
