### 2019-02-02 9ad6486
Update .trigger.sh
