### 2019-02-07 f3f54ac
Use fromJSON in jsonlite instead of rjson
