+++ DESCRIPTION
@@ -25,7 +25,7 @@ Imports:
     httr,
     assertthat,
     digest,
-    rjson,
+    jsonlite,
     data.table
 Suggests:
     testthat,
