+++ NAMESPACE
@@ -26,6 +26,6 @@ importFrom(data.table,setorder)
 importFrom(digest,digest)
 importFrom(httr,GET)
 importFrom(httr,content)
-importFrom(rjson,fromJSON)
+importFrom(jsonlite,fromJSON)
 importFrom(utils,getFromNamespace)
 importFrom(utils,packageVersion)
