### 2019-02-08 #261 @stefaniebutland
@juyeongkim following up. I know it's 💯hard to do all the work AND write a post. We would really love to help get more eyes on your work but I don't want to bug you. Let me know if you want me to pester you about scheduling a post of your choice.
