+++ README.Rmd
@@ -55,7 +55,7 @@ This will create a netrc file in your home directory.
 
 ***Alternatively***, you can manually create a netrc file in the computer running R.
 
-* On Windows, this file sould be named `_netrc`
+* On Windows, this file should be named `_netrc`
 * On UNIX, it should be named `.netrc`
 * The file should be located in the user's home directory, and the permissions on the file should be unreadable for everybody except the owner
 * To determine home directory, run `Sys.getenv("HOME")` in R
@@ -129,7 +129,7 @@ knitr::kable(cvd408$availableDatasets)
 which will print names of available datasets.
 
 
-### Neutralizing Antibody dataset (`NAb`) can be retreived by:
+### Neutralizing Antibody dataset (`NAb`) can be retrieved by:
 
 ```{r}
 NAb <- cvd408$getDataset("NAb")
@@ -139,7 +139,7 @@ colnames(NAb)
 
 Check out [the reference page](https://ropensci.github.io/DataSpaceR/reference/DataSpaceStudy.html) of `DataSpaceStudy` for all available fields and methods.
 
-***Note***: The package uses a [R6](https://cran.r-project.org/web/packages/R6/index.html) class to represent the connection to a study and get around some of R's copy-on-change behaviour.
+***Note***: The package uses a [R6](https://cran.r-project.org/web/packages/R6/index.html) class to represent the connection to a study and get around some of R's copy-on-change behavior.
 
 
 ## Examples & Documentation
