### 2019-02-12 9f25d65
Remove LICENSE file for CRAN submission
