+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: DataSpaceR
 Type: Package
 Title: An R Interface to the CAVD DataSpace
-Version: 0.6.0
+Version: 0.6.1
 Authors@R: c(person("Ju Yeong", "Kim",
                     email = "jkim2345@fredhutch.org",
                     role = c("aut", "cre")),
@@ -14,7 +14,7 @@ Description: Provides a convenient API for accessing datasets within
     the CAVD DataSpace (dataspace.cavd.org).
 URL: https://github.com/ropensci/DataSpaceR
 BugReports: https://github.com/ropensci/DataSpaceR/issues
-License: GPL-3 + file LICENSE
+License: GPL-3
 Encoding: UTF-8
 LazyData: true
 Imports:
