### 2019-02-22 692cb6e
Revise DESCRIPTION and change default path to write netrc file from home to temp
