### 2019-04-02 97fde30
Prepare for CRAN submission
